third_party/WebKit/LayoutTests/animations/keyframe-effect-options-composite.html

<!DOCTYPE html>
<title>Regression test for KeyframeEffectOptions composite value</title>

<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<div id="tester"></div>

<script>
function disableCSSAdditiveAnimationsForTest(t) {
  var cssAdditiveAnimationsEnabled = internals.runtimeFlags.cssAdditiveAnimationsEnabled;
  internals.runtimeFlags.cssAdditiveAnimationsEnabled = false;

  t.add_cleanup(() => {
    internals.runtimeFlags.cssAdditiveAnimationsEnabled = cssAdditiveAnimationsEnabled;
  });
}

function disableWebAnimationsAPIForTest(t) {
  var webAnimationsAPIEnabled = internals.runtimeFlags.webAnimationsAPIEnabled;
  internals.runtimeFlags.webAnimationsAPIEnabled = false;

  t.add_cleanup(() => {
    internals.runtimeFlags.webAnimationsAPIEnabled = webAnimationsAPIEnabled;
  });
}

// These are a set of regression tests to ensure that we don't break backwards
// compatibility for parsing the composite value of KeyframeEffectOptions.
// See http://crbug.com/806139.

test(t => {
  disableCSSAdditiveAnimationsForTest(t);
  disableWebAnimationsAPIForTest(t);

  const tester = document.getElementById('tester');
  tester.animate({ backgroundColor: [ 'red', 'blue' ] }, { composite: 'replace' });
  tester.animate({ backgroundColor: [ 'red', 'blue' ] }, { composite: 'add' });
  tester.animate({ backgroundColor: [ 'red', 'blue' ] }, { composite: 'accumulate' });

  assert_throws(new TypeError(), function() {
    tester.animate({ backgroundColor: [ 'red', 'blue' ] }, { composite: 'foo' });
  });
}, 'In stable Chrome, valid composite values should not throw');

test(t => {
  disableWebAnimationsAPIForTest(t);

  const tester = document.getElementById('tester');
  tester.animate({ backgroundColor: [ 'red', 'blue' ] }, { composite: 'replace' });
  tester.animate({ backgroundColor: [ 'red', 'blue' ] }, { composite: 'add' });
  tester.animate({ backgroundColor: [ 'red', 'blue' ] }, { composite: 'accumulate' });

  assert_throws(new TypeError(), function() {
    tester.animate({ backgroundColor: [ 'red', 'blue' ] }, { composite: 'foo' });
  });
}, 'When just CSSAdditiveAnimations is enabled, valid composite values should still not throw');

test(t => {
  disableCSSAdditiveAnimationsForTest(t);

  const tester = document.getElementById('tester');
  const anim1 = tester.animate({ backgroundColor: [ 'red', 'blue' ] }, { composite: 'replace' });
  const anim2 = tester.animate({ backgroundColor: [ 'red', 'blue' ] }, { composite: 'add' });

  assert_throws(new TypeError(), function() {
    tester.animate({ backgroundColor: [ 'red', 'blue' ] }, { composite: 'accumulate' });
  });
  assert_throws(new TypeError(), function() {
    tester.animate({ backgroundColor: [ 'red', 'blue' ] }, { composite: 'foo' });
  });
}, 'When CSSAdditiveAnimations are disabled but Web Animations are enabled, we throw on ' +
    'accumulate but not add');
</script>
